Move ControllerArgumentValueResolverPass to the HttpKernel component
