+++ UPGRADE-3.3.md
@@ -148,6 +148,10 @@ FrameworkBundle
 
  * Extending `ConstraintValidatorFactory` is deprecated and won't be supported in 4.0.
 
+ * The `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\ControllerArgumentValueResolverPass` class
+   has been removed. Use the `Symfony\Component\HttpKernel\DependencyInjection\ControllerArgumentValueResolverPass`
+   class instead.
+
 HttpKernel
 -----------
